transferred the docking implementation to DropController
