+++ src/appshell/view/dockwindow/dockwindow.cpp
@@ -40,35 +40,6 @@
 using namespace mu::dock;
 using namespace mu::async;
 
-static constexpr double MAX_DISTANCE_TO_HOLDER = 50;
-
-static bool isPointAllowedForDrop(const QPoint& point, const DropDestination& dropDestination)
-{
-    QRect dropRect = dropDestination.dock->frameGeometry();
-
-    if (!dropRect.contains(point)) {
-        return false;
-    }
-
-    if (dropDestination.dropDistance == 0) {
-        return true;
-    }
-
-    if (dropDestination.dropLocation == Location::Left) {
-        if (std::abs(dropRect.left() - point.x()) <= dropDestination.dropDistance) {
-            return true;
-        }
-    }
-
-    if (dropDestination.dropLocation == Location::Right) {
-        if (std::abs(dropRect.right() - point.x()) <= dropDestination.dropDistance) {
-            return true;
-        }
-    }
-
-    return false;
-}
-
 DockWindow::DockWindow(QQuickItem* parent)
     : QQuickItem(parent),
     m_toolBars(this),
@@ -231,39 +202,14 @@ void DockWindow::toggleDockFloating(const QString& dockName)
     }
 }
 
-DropDestination DockWindow::hover(const QString& draggedDockName, const QPoint& globalPos)
+DockPageView* DockWindow::currentPage() const
 {
-    DockBase* draggedDock = m_currentPage ? m_currentPage->dockByName(draggedDockName) : nullptr;
-
-    if (!draggedDock) {
-        return DropDestination();
-    }
-
-    QPoint hoveredLocalPos = m_mainWindow->mapFromGlobal(globalPos);
-    DropDestination dropDestination = resolveDropDestination(draggedDock, hoveredLocalPos);
-
-    if (auto toolBar = dynamic_cast<DockToolBarView*>(draggedDock)) {
-        updateToolBarOrientation(toolBar, dropDestination);
-    }
-
-    setCurrentDropDestination(draggedDock, dropDestination);
-
-    return m_currentDropDestination;
+    return m_currentPage;
 }
 
-void DockWindow::endHover()
+QQuickItem& DockWindow::asItem() const
 {
-    if (!m_currentDropDestination.isValid()) {
-        return;
-    }
-
-    m_currentDropDestination.dock->hideHighlighting();
-
-    if (m_currentDropDestination.dock->type() == DockType::DockingHolder) {
-        m_currentDropDestination.dock->close();
-    }
-
-    m_currentDropDestination.clear();
+    return *m_mainWindow;
 }
 
 DockingHolderView* DockWindow::mainToolBarDockingHolder() const
@@ -554,238 +500,4 @@ QList<DockToolBarView*> DockWindow::topLevelToolBars(const DockPageView* page) c
     }
 
     return toolBars;
-}
-
-bool DockWindow::isMouseOverDock(const QPoint& mouseLocalPos, const DockBase* dock) const
-{
-    QRect geometry = dock ? dock->frameGeometry() : QRect();
-    return geometry.contains(mouseLocalPos);
-}
-
-void DockWindow::updateToolBarOrientation(DockToolBarView* draggedToolBar, const DropDestination& dropDestination)
-{
-    IF_ASSERT_FAILED(draggedToolBar) {
-        return;
-    }
-
-    framework::Orientation orientation = framework::Orientation::Horizontal;
-
-    if (!dropDestination.isValid()) {
-        draggedToolBar->setOrientation(static_cast<Qt::Orientation>(orientation));
-        return;
-    }
-
-    switch (dropDestination.dock->location()) {
-    case Location::Left:
-    case Location::Right:
-        orientation = framework::Orientation::Vertical;
-        break;
-    case Location::Top:
-    case Location::Bottom:
-        orientation = framework::Orientation::Horizontal;
-        break;
-    case Location::Center:
-    case Location::Undefined:
-        break;
-    }
-
-    draggedToolBar->setOrientation(static_cast<Qt::Orientation>(orientation));
-}
-
-void DockWindow::setCurrentDropDestination(const DockBase* draggedDock, const DropDestination& dropDestination)
-{
-    if (m_currentDropDestination == dropDestination) {
-        return;
-    }
-
-    endHover();
-
-    m_currentDropDestination = dropDestination;
-
-    if (!m_currentDropDestination.isValid()) {
-        return;
-    }
-
-    auto showHighlighting = [this, draggedDock]() {
-        QRect highlightingRect = resolveHighlightingRect(draggedDock, m_currentDropDestination);
-        m_currentDropDestination.dock->showHighlighting(highlightingRect);
-    };
-
-    if (m_currentDropDestination.dock->type() != DockType::DockingHolder) {
-        showHighlighting();
-        return;
-    }
-
-    switch (m_currentDropDestination.dock->location()) {
-    case Location::Left:
-    case Location::Right:
-        m_currentDropDestination.dock->setMinimumWidth(draggedDock->minimumWidth());
-        break;
-    case Location::Top:
-    case Location::Bottom:
-        m_currentDropDestination.dock->setMinimumHeight(draggedDock->minimumHeight());
-        break;
-    case Location::Center:
-    case Location::Undefined:
-        break;
-    }
-
-    m_currentDropDestination.dock->open();
-    showHighlighting();
-    m_currentDropDestination.dock->init();
-}
-
-DropDestination DockWindow::resolveDropDestination(const DockBase* draggedDock, const QPoint& localPos) const
-{
-    if (draggedDock->type() == DockType::Panel) {
-        DropDestination destination;
-
-        destination.dock = resolveTabifyPanel(dynamic_cast<const DockPanelView*>(draggedDock), localPos);
-        destination.dropLocation = resolveDropLocation(destination.dock, localPos);
-
-        if (destination.isValid()) {
-            return destination;
-        }
-    }
-
-    const DockingHolderView* holder = resolveDockingHolder(draggedDock->type(), localPos);
-    QList<DropDestination> destinations = draggedDock->dropDestinations();
-
-    for (const DropDestination& destination : destinations) {
-        if (holder == destination.dock) {
-            return destination;
-        }
-
-        if (isPointAllowedForDrop(localPos, destination)) {
-            return destination;
-        }
-    }
-
-    return DropDestination();
-}
-
-DockingHolderView* DockWindow::resolveDockingHolder(DockType draggedDockType, const QPoint& localPos) const
-{
-    if (!m_currentPage) {
-        return nullptr;
-    }
-
-    if (!m_mainWindow->contains(localPos)) {
-        return nullptr;
-    }
-
-    QRect centralGeometry = m_currentPage->centralDock()->frameGeometry();
-
-    // TODO: Need to take any panels docked at top into account
-    if (localPos.y() <= centralGeometry.top() + MAX_DISTANCE_TO_HOLDER) {
-        return m_currentPage->holder(draggedDockType, Location::Top);
-    }
-
-    if (localPos.y() >= centralGeometry.bottom() - MAX_DISTANCE_TO_HOLDER) {
-        return m_currentPage->holder(draggedDockType, Location::Bottom);
-    }
-
-    if (localPos.x() <= MAX_DISTANCE_TO_HOLDER) {
-        return m_currentPage->holder(draggedDockType, Location::Left);
-    }
-
-    if (localPos.x() >= m_mainWindow->rect().right() - MAX_DISTANCE_TO_HOLDER) {
-        return m_currentPage->holder(draggedDockType, Location::Right);
-    }
-
-    return nullptr;
-}
-
-DockPanelView* DockWindow::resolveTabifyPanel(const DockPanelView* panel, const QPoint& localPos) const
-{
-    QList<DockPanelView*> tabs = m_currentPage->possibleTabs(panel);
-
-    for (DockPanelView* tab : tabs) {
-        if (isMouseOverDock(localPos, tab)) {
-            return tab;
-        }
-    }
-
-    return nullptr;
-}
-
-Location DockWindow::resolveDropLocation(const DockBase* hoveredDock, const QPoint& localPos) const
-{
-    if (!hoveredDock) {
-        return Location::Undefined;
-    }
-
-    QRect geometry = hoveredDock->frameGeometry();
-    Location dockLocation = hoveredDock->location();
-
-    qreal frameEnd = geometry.bottom();
-    qreal mousePos = localPos.y();
-    Location beginDropLocation = Location::Top;
-    Location endDropLocation = Location::Bottom;
-
-    if (dockLocation == Location::Top || dockLocation == Location::Bottom) {
-        mousePos = localPos.x();
-        frameEnd = geometry.right();
-        beginDropLocation = Location::Left;
-        endDropLocation = Location::Right;
-    }
-
-    if (mousePos <= frameEnd / 3) {
-        return beginDropLocation;
-    }
-
-    if (mousePos <= frameEnd / 1.5) {
-        return Location::Center;
-    }
-
-    if (mousePos <= frameEnd) {
-        return endDropLocation;
-    }
-
-    return Location::Undefined;
-}
-
-QRect DockWindow::resolveHighlightingRect(const DockBase* draggedDock, const DropDestination& destination) const
-{
-    if (!destination.isValid()) {
-        return QRect();
-    }
-
-    QRect frameGeometry = destination.dock->frameGeometry();
-    int frameWidth = frameGeometry.width();
-    int frameHeight = frameGeometry.height();
-    QRect fullFrameHighlightingRect = QRect(0, 0, frameWidth, frameHeight);
-
-    if (destination.dock->type() == DockType::DockingHolder) {
-        return fullFrameHighlightingRect;
-    }
-
-    if (destination.dock->type() == DockType::Central) {
-        int draggedDockWidth = draggedDock->frameGeometry().width();
-
-        if (destination.dropLocation == Location::Left) {
-            return QRect(0, 0, draggedDockWidth, frameHeight);
-        }
-
-        if (destination.dropLocation == Location::Right) {
-            return QRect(frameWidth - draggedDockWidth, 0, draggedDockWidth, frameHeight);
-        }
-    }
-
-    switch (destination.dropLocation) {
-    case Location::Top:
-        return QRect(0, 0, frameWidth, frameHeight / 2);
-    case Location::Bottom:
-        return QRect(0, frameHeight / 2, frameWidth, frameHeight / 2);
-    case Location::Left:
-        return QRect(0, 0, frameWidth / 2, frameHeight);
-    case Location::Right:
-        return QRect(frameWidth / 2, 0, frameWidth / 2, frameHeight);
-    case Location::Center:
-        return fullFrameHighlightingRect;
-    case Location::Undefined:
-        break;
-    }
-
-    return QRect();
 }
